Migrated reducers to TS
